[move] 불필요한 패키지 삭제 및 파일 이동
